feat(model): 实现SqLinear模型与数据加载器
- 新增TrafficDataset类用于加载交通数据并执行patch划分
- 实现get_dataloader函数提供数据加载接口
- 添加SpatioTemporalEmbedding时空嵌入层
- 实现HLIBlock分层线性交互模块
- 构建SqLinear完整模型结构
- 修复partition.py中split_pos类型转换问题
- 添加测试脚本验证数据加载功能

+++ core/partition.py
@@ -21,7 +21,7 @@ def separate_partitions(nodes, capacity):
     nodes.sort(key=lambda node: getattr(node, split_axis))
 
     # Step 2: Identify splitting position
-    split_pos = capacity * math.ceil(len(nodes) / capacity) // 2
+    split_pos = int(capacity * math.ceil(len(nodes) / capacity) // 2)
 
     # Step 3: Partition and recurse
     nodes1 = nodes[:split_pos]
